web-app/requirements.txt
flask
flask_login
flask_migrate
flask_babelex
WTForms==2.2.1 
flask_wtf==0.14.2
flask_sqlalchemy
flask-uploads
gunicorn
xlrd
opencage
werkzeug==0.16.0
pandas
numpy
psycopg2
nltk
Flask-Script
xlsxwriter
names
geonamescache
pycountry
dateparser==0.7.4
celery